Extract settings_bots.js.
