+++ .eslintrc.json
@@ -37,6 +37,7 @@
         "settings_notifications": false,
         "settings_muting": false,
         "settings_lab": false,
+        "settings_bots": false,
         "settings": false,
         "resize": false,
         "loading": false,
