+++ frontend_tests/node_tests/settings_bots.js
@@ -1,19 +1,15 @@
-var jsdom = require("jsdom");
-var window = jsdom.jsdom().defaultView;
-global.$ = require("jquery")(window);
-
 set_global("page_params", {
     realm_uri: "https://chat.example.com",
 });
 
-var settings = require("js/settings.js");
+var settings_bots = require("js/settings_bots.js");
 
 (function test_generate_zuliprc_uri() {
     var bot = {
         email: "error-bot@zulip.org",
         api_key: "QadL788EkiottHmukyhHgePUFHREiu8b",
     };
-    var uri = settings.generate_zuliprc_uri(bot.email, bot.api_key);
+    var uri = settings_bots.generate_zuliprc_uri(bot.email, bot.api_key);
     var expected = "data:application/octet-stream;charset=utf-8," + encodeURIComponent(
         "[api]\nemail=error-bot@zulip.org\n" +
         "key=QadL788EkiottHmukyhHgePUFHREiu8b\n" +
@@ -28,7 +24,7 @@ var settings = require("js/settings.js");
         email: "admin12@chatting.net",
         api_key: "nSlA0mUm7G42LP85lMv7syqFTzDE2q34",
     };
-    var content = settings.generate_zuliprc_content(user.email, user.api_key);
+    var content = settings_bots.generate_zuliprc_content(user.email, user.api_key);
     var expected = "[api]\nemail=admin12@chatting.net\n" +
                    "key=nSlA0mUm7G42LP85lMv7syqFTzDE2q34\n" +
                    "site=https://chat.example.com\n";
